refactor commit reset menu
